Move mock llm functions to their own file
